+++ src/components/ToastContainer/Toast/index.tsx
@@ -1,0 +1,28 @@
+import React from 'react';
+import { FiAlertCircle, FiXCircle } from 'react-icons/fi';
+import { ToastMessage, useToast } from '../../../hooks/toast';
+
+import { Container } from './styles';
+
+interface ToastProps {
+  message: ToastMessage;
+}
+
+const Toast: React.FC<ToastProps> = ({ message }) => {
+  const { removeToast } = useToast();
+
+  return (
+    <Container type={message.type} hasDescription={!!message.description}>
+      <FiAlertCircle size={20} />
+      <div>
+        <strong>{message.title}</strong>
+        {message.description && <p>{message.description}</p>}
+        <button onClick={() => removeToast(message.id)} type="button">
+          <FiXCircle size={18} />
+        </button>
+      </div>
+    </Container>
+  );
+};
+
+export default Toast;
