+++ src/components/ToastContainer/index.tsx
@@ -1,32 +1,18 @@
 import React from 'react';
-import { FiAlertCircle, FiXCircle } from 'react-icons/fi';
-import { ToastMessage, useToast } from '../../hooks/toast';
-import { Conatainer, Toast } from './styles';
+import { ToastMessage } from '../../hooks/toast';
+import Toast from './Toast';
+
+import { Conatainer } from './styles';
 
 interface ToastContainerProps {
   messages: ToastMessage[];
 }
 
 const ToastContainer: React.FC<ToastContainerProps> = ({ messages }) => {
-  const { removeToast } = useToast();
-
   return (
     <Conatainer>
       {messages.map(message => (
-        <Toast
-          key={message.id}
-          type={message.type}
-          hasDescription={!!message.description}
-        >
-          <FiAlertCircle size={20} />
-          <div>
-            <strong>{message.title}</strong>
-            {message.description && <p>{message.description}</p>}
-            <button onClick={() => removeToast(message.id)} type="button">
-              <FiXCircle size={18} />
-            </button>
-          </div>
-        </Toast>
+        <Toast key={message.id} message={message} />
       ))}
     </Conatainer>
   );
